Refactor the mission script as a finite state automaton
The mission script is fully refactored and the logic is altered in a lot of places. The script now consists of several `state` functions each handling different game states. The algorithm is relatively stable and seems to handle most situations.

The new algorithm also takes advantage of multiple threads, notably - two threads always exist, one of which is a capture worker that is fully responsible for screen capture at a maximum FPS